ipc plugins: Move list-views from stipc to ipc-rules
This also lists the views in z-order.
